refactor(detection): remove the structurally-null 'detection lead' metric; keep MTTD + RTT-degradation (closes #679)
The "detection lead" (faster-than-official) metric was structurally dead: status-page
polling is always LATER than the official publish, so the lead was permanently negative
(never displayed). #464 already retired the "faster than official" framing; production KV
re-confirmed in_window=0 across every recorded incident. Removed as dead complexity.

Removed:
- `formatDetectionLead` (alerts.ts) + the whole `detection-lead-log.ts` module
  (`computeLeadMs`/`appendDetectionLead`/`readDetectionLeadEntries`/`formatDetectionLeadSection`
  + the diag: `classifyLead`/`appendLeadDiag`/`readLeadDiag`/`formatLeadDiagSection`/`LeadDiag`
  + `DetectionLeadEntry`/`MIN_LEAD_SAMPLE_SIZE` + the `detection:lead:{date}`/`:monthly`/`:diag` KV).
- `MonthlyArchive.detectionLead` + `MonthlyDetectionLeadSummary`/`summarizeDetectionLead`/
  `canPresentLeadAverage` (monthly-archive.ts) — `/api/report` no longer carries detectionLead.
- The per-incident lead embed line + the daily-summary lead/diag reads + sections + data fields.

diag decision = DROP: `classifyLead`/the diag buckets only measured the now-removed lead
(detectedAt vs officialAt); the "detection works" insight is documented, and MTTD derives from
the cron cadence, not the diag.

Kept (intact): `detected:{svcId}` (reused by #677 + the MTTD framing) · `detection.ts` (the
`detected:` capture) · RTT-degradation detection (`probe-degradation:*`, `MonthlyArchive.degradation`,
`summarizeDegradation`, `formatDegradationSection`, degradationCounts). `classifyDegradation` +
`DegradationOutcome` MOVED from the deleted detection-lead-log.ts into daily-summary.ts (next to
`formatDegradationSection`), behavior identical; its single index.ts call site (RTT spike loop) intact.

Cross-repo: aiwatch-reports `generate-report.js` reads `archive.detectionLead` but GUARDS it
(conditional render), so the worker omitting it is graceful (renders nothing for new months) — a
reports-side cleanup is a separate follow-up.

Tests: removed the lead/diag tests, moved the classifyDegradation test into daily-summary.test.ts,
dropped the orphaned monthly-archive lead fixtures. test:worker 1669 · typecheck:worker clean ·
wrangler --dry-run green. Docs: kv-schema.md (retire the 3 lead/diag rows, keep `detected:`),
CLAUDE.md, README(.ko), data-flow.md all updated. Worker/Discord-only — no dashboard/is-down surface.

Worker deploy required after merge.
