+++ docs/reference/data-flow.md
@@ -22,15 +22,15 @@ Cron Trigger (*/5 min)
   → platform monitor: check metastatuspage.com → store platform:status:atlassian → Discord alert on outage/recovery
   → read KV cache → detect incidents/status changes
   → record detection timestamps (detected:{serviceId}); probe-spike rising edge also increments probe-degradation:daily(+:nostatus) RTT-degradation counters (#464)
-  → KV ID-based dedup → Discord alerts (single embed per incident; rare genuine early-RTT signal appended when probe flagged degradation before the official update)
+  → KV ID-based dedup → Discord alerts (single embed per incident)
   → #633 first-seen confirmation gate: a flap-shaped NEW incident (`isFlapNotice`: "— down/recovered" title + impact **not `major`** — #564/#565 maps a BetterStack monitor flap to `minor`/null, so the gate excludes only explicit broad-outage `major`, not all non-null impact; the original `impact==null` check silently disabled the gate post-#565 → Modal phantom recurrence) on a `flapSuppression` monitor-flap service (together/fireworks/huggingface/modal/luma) is HELD for one cron cycle — alert + AI analysis fire only after it survives a prior cycle (`pending:new:{incId}`, 600s TTL). A single-cycle BetterStack blip that self-recovers inside the window never alerts (no phantom). Severity-tagged incidents + Tier-1 (claude/openai/gemini) are never held → immediate alert. Symmetric with the degraded-debounce below
-  → incident detected → AI analysis via Gemma 4 (Workers AI, primary) or Sonnet (AI Gateway, fallback) (8s timeout) + early-RTT signal (1-60min, rare → "⚡ Early signal: Xm") → merged into incident embed + persisted to detection:lead:{date} audit log (#256, dedup by incId). NOTE (#464): status-page polling is structurally later than the official publish, so the "faster than official" headline is retired — honest framing is MTTD (alerted within ~5-min polling cycle) + RTT degradation detection (degradations status pages don't report); aggregate average gated by MIN_LEAD_SAMPLE_SIZE
+  → incident detected → AI analysis via Gemma 4 (Workers AI, primary) or Sonnet (AI Gateway, fallback) (8s timeout). Detection framing (#464/#679): status-page polling is structurally later than the official publish, so the "faster than official" lead metric was removed — honest framing is MTTD (alerted within ~5-min polling cycle) + RTT degradation detection (degradations the status pages don't report)
   → recovery detected → mark ai:analysis:{svcId}:{incId} with resolvedAt (2h TTL, powers "Recently Resolved" UI)
   → active incidents: refresh analysis TTL / re-analyze if expired / dedup sibling services
   → alert count tracked in KV (alert:count:{date}) for Daily Summary
-  → daily summary at UTC 09:00 (KST 18:00) with alert count aggregation + Web Vitals p75 + early-RTT detections (24h sliding window) + RTT degradation counts (probe-degradation:daily, #464) + fetch-failure observability (#500/#501)
+  → daily summary at UTC 09:00 (KST 18:00) with alert count aggregation + Web Vitals p75 + RTT degradation counts (probe-degradation:daily, #464) + fetch-failure observability (#500/#501)
   → daily summary also accumulates incidents:monthly:{YYYY-MM} (dedup by incident ID, 60d TTL)
-  → monthly archive on 1st of month (UTC 00:00) → aggregate history:* + probe:daily:* + incidents:monthly:* + security:monthly:* + detection:lead:monthly:* (#369) + probe-degradation:monthly:* (#511) → archive:monthly:{YYYY-MM} (permanent)
+  → monthly archive on 1st of month (UTC 00:00) → aggregate history:* + probe:daily:* + incidents:monthly:* + security:monthly:* + probe-degradation:monthly:* (#511) → archive:monthly:{YYYY-MM} (permanent)
   → archive-ready Discord ping → links to `aiwatch-reports/generate-report.yml` workflow_dispatch so operator clicks "Run workflow" to open draft PR; dedup via archive:notified:{YYYY-MM} (aiwatch-reports#4)
   → changelog RSS/HTML collection (hourly at :00) → KV accumulate new entries from OpenAI/Google/Anthropic
   → security monitoring (hourly at :00) → HN Algolia + OSV.dev SDK vulnerability scan (24 AI SDK packages · two-phase: querybatch → KV dedup → per-vuln GET enrichment; #323/#325) → EPSS enrichment via GitHub Advisories (24h KV cache, #326) → Discord digest on findings
