+++ docs/reference/kv-schema.md
@@ -22,13 +22,10 @@
 | `alerted:flap:{svcId}:{normalizedTitle}` | `"1"` | 1h | ~5-20 | BetterStack auto-recovery flap suppression (#283) — 60-min window per service + title. Written on the `alerted:res:` fire of the first flap; checked on next cron cycle to drop both down + resolved halves of subsequent identical flaps. Opt-in via `ServiceConfig.flapSuppression: true`; applies to `together`, `fireworks`, `huggingface`, `modal`, `luma`. Tier-1 (`claude`/`openai`/`gemini`) never suppressed as defense-in-depth |
 | `pending:degraded:{svcId}` | `"1"` | 10min | ~5 | Anti-flapping: 2-cycle consecutive detection |
 | `pending:new:{incId}` | `"1"` | 10min | ~0-5 | First-seen confirmation gate (#633) — written when a flap-shaped NEW incident (`isFlapNotice`: "— down/recovered" title + impact **not `major`**; #564/#565 maps a BetterStack monitor flap to `minor`/null, so the gate excludes only `major`, not all non-null impact) on a `flapSuppression` service (together/fireworks/huggingface/modal/luma) is HELD on its first cron sighting. The next cycle reads it to confirm + fire the alert; a single-cycle BetterStack blip that self-recovers lets the marker expire → no phantom alert. Mirrors `pending:degraded` on the new-incident path |
-| `detected:{svcId}` | ISO timestamp | 7d | ~5 | Detection Lead: earliest detection time (probe spike or status page, whichever is earlier) |
-| `detection:lead:{YYYY-MM-DD}` | `DetectionLeadEntry[]` JSON | 7d | ~0-5 | Detection Lead audit log — appended on each new incident with positive lead, dedup by incId, surfaced in Daily Summary Discord embed (#256) |
-| `detection:lead:monthly:{YYYY-MM}` | `DetectionLeadEntry[]` JSON | 60d | ~0-5 | Detection Lead monthly accumulator (#369) — dual-written by `appendDetectionLead` alongside the daily key so the monthly archive cron can read full-month entries past the 7d daily TTL. Summarized into `MonthlyArchive.detectionLead` (count / avg / median / max / byService / topExamples). 60d TTL covers archive's 1st-of-next-month read window with margin |
-| `detection:lead:diag:{YYYY-MM-DD}` | `LeadDiag` JSON `{probe,nonProbe:{no_detected,negative,below_min,in_window,above_max}}` | 30d | ~0-N | Detection Lead **diagnostics** counter (#464) — lossy daily tally (not source-of-truth) of *why* each new incident did/didn't produce a lead, split by probe coverage. Incremented by `appendLeadDiag` in the new-incident alert path; surfaced as a Daily Summary line via `formatLeadDiagSection`. Measures whether the empty audit log is a coverage vs timing problem. **#464 redefinition**: diag data showed `in_window`=0 (status-page polling is structurally later than the official publish), so the "faster than official" framing is retired — the aggregate "average lead" claim is gated by `MIN_LEAD_SAMPLE_SIZE` (consumer-side `canPresentLeadAverage`), and the honest framing is MTTD ("alerted within ~5-min polling cycle") + RTT degradation detection below |
+| `detected:{svcId}` | ISO timestamp | 7d | ~5 | Earliest detection time (probe spike or status page, whichever is earlier). Feeds the MTTD framing + reused by #677 as the AWS/Bedrock incident-start anchor. **#679** — the per-incident "detection lead" (faster-than-official) surfacing + the `detection:lead:{date}`/`:monthly`/`:diag` accumulators were REMOVED (the lead was structurally null — status-page polling is always later than the official publish; #464 already retired the framing). This key itself is KEPT |
 | `probe-degradation:daily:{svcId}:{YYYY-MM-DD}` | counter string | 48h | ~0-N | RTT degradation counter (#464) — incremented on the **rising edge** of a probe-spike streak (gated by the existing `alerted:probe-spike:` 1h dedup). The honest differentiator: AIWatch's direct RTT probe catches latency degradation that status pages often never report. Read at the daily summary via `formatDegradationSection` |
 | `probe-degradation:nostatus:daily:{svcId}:{YYYY-MM-DD}` | counter string | 48h | ~0-N | Subset of `probe-degradation:daily` where the service's official status was still `operational` at spike time — i.e. the degradation is NOT on the status page. The headline figure for "degradations status pages never report" (#464). Classified by `classifyDegradation` against the cached `services:latest` snapshot |
-| `probe-degradation:monthly:{YYYY-MM}` | `DegradationMonthly` JSON `{byService,noStatusByService}` | 60d | ~0-N | RTT degradation monthly accumulator (#511) — dual-written by the spike loop alongside the daily counters (read-modify-write via `addDegradationToMonthly`) so the monthly archive cron can read month-complete counts past the daily 48h TTL. Summarized into `MonthlyArchive.degradation` (total / noStatusTotal / byService / noStatusByService) and exposed via `/api/report`. Mirrors `detection:lead:monthly` |
+| `probe-degradation:monthly:{YYYY-MM}` | `DegradationMonthly` JSON `{byService,noStatusByService}` | 60d | ~0-N | RTT degradation monthly accumulator (#511) — dual-written by the spike loop alongside the daily counters (read-modify-write via `addDegradationToMonthly`) so the monthly archive cron can read month-complete counts past the daily 48h TTL. Summarized into `MonthlyArchive.degradation` (total / noStatusTotal / byService / noStatusByService) and exposed via `/api/report`. Mirrors the `security:monthly` accumulator pattern |
 | `reddit:seen:{postId}` | `"1"` | 24h | ~120 | Reddit post dedup (hourly scan, max 5/hour) |
 | `security:seen:hn:{objectId}` | `SecurityAlertMeta` JSON | 7d | ~0 | HN security post dedup + dashboard display |
 | `security:seen:osv:{vulnId}` | `SecurityAlertMeta` JSON | 7d | ~0 | OSV.dev vulnerability dedup + dashboard display (includes `epssPercentile` / `epssPercentage` from #326 when GitHub Advisories enrichment succeeded) |
